backend/requirements/base.txt
Django==3.2.20
django-model-utils==4.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
django-cors-headers==3.13.0
django-filter==2.4.0
raven==6.0.0
psycopg2==2.8.4
sentry-sdk==1.14.0
rest-condition==1.0.3
xlrd==1.2.0
Pillow==9.3.0
XlsxWriter==1.3.7
drf-renderer-xlsx==0.4.1
requests==2.31.0
setuptools>=65.5.1
